Move styles to stylesheet folder
